feat(add-data): Service list component
