+++ projects/hslayers/src/components/add-data/common/url/layer-table/layer-table.component.html
@@ -1,0 +1,119 @@
+<p>layer-table works!</p>
+<div>
+    <ng-container *ngIf="type === 'wms'; else simpleTable">
+        <table class="table table-sm table-striped table-bordered hs-url-table">
+            <thead class="bg-secondary text-light">
+                <th style="width:1.5em">&nbsp;</th>
+                <th>{{'COMMON.title' | translateHs: {app} }}</th>
+                <!-- <th style="width:30%">{{'COMMON.name' | translateHs: {app} }}</th> -->
+                <th>{{'COMMON.abstract' | translateHs: {app} }}</th>
+            </thead>
+            <tbody *ngFor="let layer of data.layers | slice:0:limitShown; trackBy:( 'Name' | trackByProperty )">
+                <tr>
+                    <td [attr.rowspan]="layer.Style?.length ? 3 : 2"
+                        [ngClass]="{'align-middle': !hasNestedLayers(layer)}">
+                        <input type="checkbox" class="checkbox-lg" [(ngModel)]="layer.checked" name="layer"
+                            style="cursor: pointer;" (change)="searchForChecked(layer)" />
+                    </td>
+                    <td hsWmsLayerHighlight style="max-width: 25.375em" class="text-truncate"
+                        [title]="'Title: ' + layer.Title  + '\nName: ' + layer.Name"
+                        (click)="hsUrlWmsService.expandTableRow($event)">{{layer.Title}}</td>
+                    <!-- <td style="max-width: 13.75em" class="text-truncate" [title]="layer.Name">{{layer.Name}}</td> -->
+                    <td hsWmsLayerHighlight style="max-width: 10.875em; width:30%" class="text-truncate"
+                        [title]="layer.Abstract" (click)="hsUrlWmsService.expandTableRow($event)">
+                        {{layer.Abstract}}</td>
+                </tr>
+                <tr *ngIf="layer.Style?.length > 1 && layer.checked">
+                    <td colspan="3">
+                        <div class="d-flex align-items-baseline px-2 w-100">
+                            <div class="d-flex align-items-baseline" style="flex-grow: 1;">
+                                <label class="control-label">{{'COMMON.style' | translateHs: {app} }}:</label>
+                                <select class="form-control form-select-sm form-select"
+                                    [(ngModel)]="layer.styleSelected" [ngModelOptions]="{standalone: true}">
+                                    <option *ngFor="let style of layer.Style" [ngValue]="style">{{style.Name}}
+                                    </option>
+                                </select>
+                            </div>
+                        </div>
+                    </td>
+                </tr>
+                <tr [hidden]="!hasNestedLayers(layer)">
+                    <td colspan="3" style="padding-left: 1em">
+                        <hs-nested-layers-table style="display: block;" [app]="app" [layers]="layer.Layer"
+                            (layerChecked)="searchForChecked($event)">
+                        </hs-nested-layers-table>
+                    </td>
+                </tr>
+            </tbody>
+        </table>
+    </ng-container>
+    <ng-template #simpleTable>
+        <table class="table table-sm table-striped table-bordered">
+            <thead>
+                <th style="width:1em">&nbsp;</th>
+                <th>{{'COMMON.name' | translateHs: {app} }}</th>
+            </thead>
+            <tbody *ngFor="let service of data.services; trackBy:( 'id' | trackByProperty )">
+                <tr>
+                    <td>
+                        <input class="checkbox-lg" type="checkbox" name="service" (change)="changed('services')"
+                            style="cursor: pointer;" [(ngModel)]="service.checked" />
+                    </td>
+                    <td (click)="expandService(service)" data-toggle="tooltip"
+                        [title]="'ADDDATA.requestServiceLayers' | translateHs: {app} " style="cursor: pointer;">
+                        {{service.name}}
+                        <span style="font-size: x-small;"><i class="icon-chevron-right"></i></span>
+                    </td>
+                </tr>
+            </tbody>
+            <tbody *ngFor="let layer of data.layers; trackBy:( data.table.trackBy | trackByProperty )">
+                <tr>
+                    <td><input class="checkbox-lg" type="checkbox" name="layer" style="cursor: pointer;"
+                            (change)="changed('layers')" [(ngModel)]="layer.checked" /></td>
+                    <td hsWmsLayerHighlight (click)="hsUrlWmsService.expandTableRow($event)" class="text-truncate"
+                        style="max-width: 45ch;">{{layer[data.table.nameProperty]}}
+                        <div *ngIf="layer.checked && type === 'wmts'"
+                            class="d-flex flex-row justify-content-end align-items-baseline mt-1">
+                            {{'ADDDATA.CATALOGUE.addAs' | translateHs: {app} }}
+                            <div class="btn-group ms-2">
+                                <button type="button" class="btn btn-sm btn-light  btn-outline-secondary"
+                                    (click)="layer.base = true" [ngClass]="{'active':layer.base}">
+                                    {{'COMPOSITONKEYWORDS.Basemap' | translateHs: {app} }}
+                                </button>
+                                <button type="button" class="btn btn-sm btn-light  btn-outline-secondary"
+                                    (click)="layer.base = false" [ngClass]="{'active': !layer.base}">
+                                    {{'COMMON.layer' | translateHs: {app} }}
+                                </button>
+                            </div>
+                        </div>
+                        <div class="alert alert-primary d-flex m-1" role="alert"
+                            *ngIf="layer.featureCount > 1000 && layer.checked && type === 'wfs'">
+                            <div class="w-100">
+                                <p class="mb-0" style="white-space: normal" [hidden]="!layer.limitFeatureCount">
+                                    {{'ADDLAYERS.WFS.selectedLayerContains' |
+                                    translateHs: {app} }}
+                                    {{layer.featureCount}} {{'ADDLAYERS.WFS.loadingToManyAtOnce' | translateHs: {app} }}
+                                </p>
+                            </div>
+                        </div>
+                    </td>
+                </tr>
+            </tbody>
+        </table>
+    </ng-template>
+    <ng-container *ngIf="type === 'wms'">
+        <div *ngIf="reachedLimit()" class="d-flex justify-content-around align-items-center m-3"><span>
+                {{getLimitTextTranslation()}}</span>
+            <div class="btn-group ms-2">
+                <button type="button" class="btn btn-sm btn-light  btn-outline-secondary"
+                    (click)="limitShown = limitShown + 100">
+                    {{'ADDDATA.CATALOGUE.showNext100' | translateHs: {app} }}
+                </button>
+                <button type="button" class="btn btn-sm btn-light  btn-outline-secondary"
+                    (click)="limitShown = data.layers.length">
+                    {{'ADDDATA.CATALOGUE.showAll' | translateHs: {app} }}
+                </button>
+            </div>
+        </div>
+    </ng-container>
+</div>
