+++ projects/hslayers/src/components/add-data/url/wfs/wfs.service.ts
@@ -38,6 +38,10 @@ class HsUrlWfsParams {
       srss: [],
       title: '',
       version: '',
+      table: {
+        trackBy: 'Name',
+        nameProperty: 'Title',
+      },
     };
     this.definedProjections = [
       'EPSG:3857',
