+++ projects/hslayers/src/components/add-data/url/wmts/wmts.service.ts
@@ -26,6 +26,10 @@ class HsUrlWmtsParams {
       layers: [],
       title: '',
       version: '',
+      table: {
+        trackBy: 'Identifier',
+        nameProperty: 'Title',
+      },
     };
   }
 }
